+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added unit test for whitespace handling.
 - Added optional argument `$trim = true` to `$node->text()`
 ### Changed
+- `$node->text()` now supports all block and inline level elements.
 ### Deprecated
 ### Removed
 ### Fixed